refactor(orders): duck/feature folder layout
Reorganizes the now-Effect orders service into the duck/feature layout from the
house-style guide, with history-preserving renames:

- domain/   — errors, state-machine, schema (pure domain + drizzle tables)
- runtime/  — config, services (tags+layers), runtime (ManagedRuntime+OTLP),
              test-runtime
- http/     — router (+ createOrderProgram), app, boundary, oRPC tests
- consumers/— order-expired, payment-created, tickets-replica (.consumer.ts)
- top-level — index, tickets-client

Relative imports, package.json exports (keys unchanged, targets repointed),
drizzle.config schema path, and the import.meta.url-relative migration paths in
the moved integration/e2e tests are all updated. No behavior change: tsgo, lint,
format, the 62 orders tests (incl. all integration suites under docker), and the
gateway canary build all pass.
